+++ src/components/ComponentName.jsx
@@ -3,16 +3,16 @@ import AppContexts from '../context/AppContext'
 
 const ComponentName = ({ result }) => {
 
-    const { addToNameState } = useContext(AppContexts);
+    const { addToNameState } = useContext(AppContexts); // example addState
 
-	const handleClick = (resultItem) => {
+	const handleClick = (resultItem) => { // example addState
 		addToNameState(resultItem)
 	}
 
     return (
         <div>
             <img src={result.img[0]} alt={result.title} />
-            <button onClick={ () => handleClick(result)}></button>
+            <button onClick={ () => handleClick(result)}></button> // example addState
         </div>
     );
 };