+++ src/containers/NameContainer.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import ComponentName from '../components/ComponentName'
-import useCrudNameAction from '../hooks/useCrudNameAction'
+import useCrudNameAction from '../hooks/useCrudNameAction' // example api call
 
-const API = '';
+const API = ''; // example api call
 
 const NameContainer = () => {
 
-    const results = useCrudNameAction(API)
+    const results = useCrudNameAction(API); // example api call
 
     return (
         <div>
             { results.map( result => (
-            <ProductItem result={result} key={result.id}/>
+            <ComponentName result={result} key={result.id}/> // example use of api
         ))}
         </div>
     );