+++ src/routes/App.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Layout from '../containers/Layout';
 import Home from '../pages/Home';
+import NotFound from '../pages/NotFound';
 import AppContext from '../context/AppContext'
 import useInitialState from '../hooks/useInitialState'
 import '../styles/resetStyle.scss'
@@ -14,7 +15,7 @@ const App = () => {
 				<Layout>
 					<Routes>
 						<Route exact path="/" element={<Home />} />
-						<Route path="*" element={<Route />} />
+						<Route path="*" element={<NotFound />} /> 
 					</Routes>
 				</Layout>
 			</BrowserRouter>